terraform: goodbye context.go (old)
